Move create_person and create_student to create_objects.rb
